Added support for debug register , through socket_dma

+++ software/network/socket_dma.h
@@ -16,7 +16,7 @@
 
 class SocketDMA {
 	static void dmaThread(void *a);
-	static void performCommand(int socket, void *load_buffer, int length, uint16_t cmd, uint16_t len);
+	static void performCommand(int socket, void *load_buffer, int length, uint16_t cmd, uint32_t len);
 	static int  readSocket(int socket, void *buffer, int max_remain);
 	static int  writeSocket(int socket, void *buffer, int length);
